# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MyStocks API
  version: 2.1.0
  description: |
    MyStocks 专业量化交易平台 API 规范。
    包含核心行情数据、技术分析、V3.1 风险管理系统及 Phase 2 信号监控模块。
  contact:
    name: 架构组
    email: arch@mystocks.com
servers:
  - url: http://localhost:8000
    description: 本地开发服务器
...
paths:
...
  /api/signals/statistics:
    get:
      tags: [SignalMonitoring]
      summary: 获取小时级信号统计
      parameters:
        - name: strategy_id
          in: query
          schema: { type: string }
        - name: hours
          in: query
          schema: { type: integer, default: 24 }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignalStatsList'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SignalStatsList:
      type: array
      items:
        type: object
        properties:
          hour: { type: string, format: "date-time" }
          signal_count: { type: integer }
          success_rate: { type: number }
          avg_profit: { type: number }
...